requirements.txt
asgiref==3.5.2
autopep8==1.7.0
backports.zoneinfo==0.2.1
Brotli==1.0.9
dj-database-url==1.0.0
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
gunicorn==20.1.0
psycopg2-binary==2.9.4
pycodestyle==2.9.1
PyJWT==2.5.0
pytz==2022.4
sqlparse==0.4.3
toml==0.10.2
whitenoise==6.2.0
